+++ src/vue/forms/TransferForm.vue
@@ -28,12 +28,19 @@
             <div class="app__form-field">
               <select-field
                 name="transfer-asset"
-                :values="transferableBalancesAssets"
-                v-model="form.asset"
-                key-as-value-text="nameAndCode"
+                :value="form.asset.code"
+                @input="setAsset"
                 :label="'transfer-form.asset-lbl' | globalize"
                 :disabled="view.mode === VIEW_MODES.confirm"
-              />
+              >
+                <option
+                  v-for="asset in transferableBalancesAssets"
+                  :key="asset.code"
+                  :value="asset.code"
+                >
+                  {{ asset.nameAndCode }}
+                </option>
+              </select-field>
               <template v-if="form.asset.code">
                 <p class="app__form-field-description">
                   {{
@@ -56,6 +63,7 @@
                 validation-type="outgoing"
                 :label="'transfer-form.amount-lbl' | globalize"
                 :asset="form.asset"
+                is-max-button-shown
               />
             </div>
           </div>
@@ -321,10 +329,10 @@ export default {
       this.isFeesLoaded = false
       setTimeout(() => this.updateView(VIEW_MODES.submit, {}, true), 1)
     },
-    setAsset () {
-      this.form.asset =
-        this.transferableBalancesAssets
-          .find(asset => asset.code === this.assetToTransfer) ||
+    setAsset (payload) {
+      const assetCode = payload || this.assetToTransfer
+      this.form.asset = this.transferableBalancesAssets
+        .find(asset => asset.code === assetCode) ||
         this.transferableBalancesAssets[0] ||
         {}
     },