+++ src/vue/modules/buy-back-form/index.vue
@@ -8,12 +8,19 @@
         <div class="app__form-row">
           <div class="app__form-field">
             <select-field
-              v-model="form.asset"
-              :values="assetsInBalance"
-              key-as-value-text="nameAndCode"
+              :value="form.asset.code"
+              @input="setAssetByCode"
               class="app__select"
               :label="'buy-back-form.opportunity-uniq-code-lbl' | globalize"
-            />
+            >
+              <option
+                v-for="asset in assetsInBalance"
+                :key="asset.code"
+                :value="asset.code"
+              >
+                {{ asset.nameAndCode }}
+              </option>
+            </select-field>
           </div>
         </div>
         <p class="app__form-field-description">
@@ -222,6 +229,10 @@ export default {
       loadAssets: types.LOAD_ASSETS,
       loadSaleByBaseAsset: types.LOAD_SALE_BY_BASE_ASSET,
     }),
+    setAssetByCode (code) {
+      this.form.asset = this.assetsInBalance
+        .find(item => item.code === code)
+    },
     async submit () {
       this.disableForm()
       this.isSubmitting = true