+++ src/vue/modules/issuance-form/components/create-issuance-form.vue
@@ -7,14 +7,21 @@
     <div class="app__form-row">
       <div class="app__form-field">
         <select-field
-          v-model="form.asset"
-          :values="ownedAssets"
+          :value="form.asset.code"
+          @input="setAssetByCode"
           name="create-issuance-asset"
-          key-as-value-text="nameAndCode"
           :label="'issuance-form.asset-lbl' | globalize"
           @blur="touchField('form.asset')"
           :disabled="formMixin.isDisabled"
-        />
+        >
+          <option
+            v-for="asset in ownedAssets"
+            :key="asset.code"
+            :value="asset.code"
+          >
+            {{ asset.nameAndCode }}
+          </option>
+        </select-field>
       </div>
     </div>
 
@@ -262,6 +269,10 @@ export default {
   },
 
   methods: {
+    setAssetByCode (code) {
+      this.form.asset = this.ownedAssets
+        .find(item => item.code === code)
+    },
     tryLoadFees () {
       this.isFeesLoaded = false
       this.isFeesLoadFailed = false