+++ src/vue/modules/withdrawal-fiat-bank/index.vue
@@ -8,14 +8,21 @@
       <div class="app__form-field">
         <select-field
           name="withdrawal-fiat-bank-asset"
-          v-model="form.asset"
-          :values="withdrawableFiatAssets"
-          key-as-value-text="nameAndCode"
+          :value="form.asset.code"
+          @input="setAssetByCode"
           :disabled="formMixin.isDisabled"
           @blur="touchField('form.asset')"
           :error-message="getFieldErrorMessage('form.asset')"
           :label="'withdrawal-fiat-bank-module.asset' | globalize"
-        />
+        >
+          <option
+            v-for="asset in withdrawableFiatAssets"
+            :key="asset.code"
+            :value="asset.code"
+          >
+            {{ asset.nameAndCode }}
+          </option>
+        </select-field>
         <div class="withdrawal-fiat-bank__form-field-description">
           <p>
             {{
@@ -316,6 +323,10 @@ export default {
       loadAssets: types.LOAD_ASSETS,
       loadFees: types.LOAD_FEES,
     }),
+    setAssetByCode (code) {
+      this.form.asset = this.withdrawableFiatAssets
+        .find(item => item.code === code)
+    },
     async getFees () {
       try {
         await this.loadFees({