+++ src/vue/pages/sales/SalesList.vue
@@ -2,13 +2,18 @@
   <div class="sales-list">
     <div class="sales__state-filter">
       <select-field
-        :is-value-translatable="true"
         :disabled="!isLoaded"
         v-model="filters.state"
-        :values="Object.values(SALE_STATES)"
-        key-as-value-text="labelTranslationId"
         class="sales-asset-selector__field app__select app__select--no-border"
-      />
+      >
+        <option
+          v-for="saleState in Object.values(SALE_STATES)"
+          :key="saleState.value"
+          :value="saleState.value"
+        >
+          {{ saleState.labelTranslationId | globalize }}
+        </option>
+      </select-field>
     </div>
     <template v-if="filteredSales.length">
       <div class="sales__sale-cards">
@@ -110,7 +115,7 @@ export default {
     saleRecords: [],
     filters: {
       baseAsset: '',
-      state: SALE_STATES.live,
+      state: SALE_STATES.live.value,
     },
     isLoaded: false,
     isDetailsDrawerShown: false,
@@ -139,7 +144,7 @@ export default {
     },
 
     recordsLoader () {
-      const saleState = this.filters.state.value
+      const saleState = this.filters.state
 
       let opts = {
         page: { order: 'desc' },